+++ app/ui/src/main/java/module-info.java
@@ -14,7 +14,10 @@ module org.pcsoft.app.aighost.ui {
 
     opens org.pcsoft.app.aighost.app to javafx.fxml, de.saxsys.mvvmfx;
     opens org.pcsoft.app.aighost.app.ui.window to javafx.fxml, de.saxsys.mvvmfx;
+    opens org.pcsoft.app.aighost.app.ui.dialog to javafx.fxml, de.saxsys.mvvmfx;
 
     exports org.pcsoft.app.aighost.app;
+    exports org.pcsoft.app.aighost.app.ui;
     exports org.pcsoft.app.aighost.app.ui.window;
+    exports org.pcsoft.app.aighost.app.ui.dialog;
 }